don't log anything in constructors
It is causing a side-effect in chat where the authentication doesn't
work anymore when not having the keys because the error is logged to
STDOUT and the ejabberd daemon needs to read STDOUT and it is choking on
the log error. This didn't happen when the keys were there but now that
some components only have the public key they were complaining about not
having the private key.