[Fix] Renew card gets card for the user in the correct tenant
· Renew card queries the principal in mongo to retrieve their card.
With multitenant now active we need to query by the principalId and
domain, otherwise we may return the wrong user if their is a collision
between users with the same username in separate domains.  This bug is
detrimental because with the wrong user card the user can eseentially
logged in as a completely different user without them knowing.

· Fixed tests